Add elements and content for intro div
This division will have a background image and a banner that is
displayed horizontally across that background image about one third
down the page. The banner will hold an icon image and the word 'ODIN',
at the bottom of the parent division we'll give credit to the
photographer in a span element.