Add flex display to body, adjust body height width
We set the body to use flex display, and set its with to use all of the
viewport width, as well as it's height to all of the viewports height.
On ios this can cause issues but I won't worry about it in this project
as the prompt says to just build this for desktop and not make it too
responsive just yet.