Add/Adjust button styling
Removed the padding from the button, it needed margin, added margin to
space it accordingly with the rest of the elements in the
content-container.
Adjusted the width of the button, making it smaller.
Added a padding to the button as the width was influencing it more than
it should.
Adjusted the font-size of the button.
Gave a border radius to make the button look calmer?
Changed the background-color and font color of the button.