Adjust form, label, and anchor elements
Changed the color of the anchor elements which navigates users to a
login page if they have already have an account, as well as removed the
underline from that text.
Added flex to the from and set up wrapping ability.
Also adjusted the form width to 80% of the available space to it, and
added a margin-top of 1rem to space it away from that h3 a bit.
Have temporarily set a width of 50% on the label elements, as I am
figuring out how to mimic the rubric given to me.