Add styling for inputs in multiple states
By default input's will have a grey border, however when a password
type input is invalid that border will have a red outline.
When an input is focused it's outline will be blue and the input will
have a box-shadow on it.
A border radius of 3px has also been applied to round the input
elements.